requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.3.4
black==24.8.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
coverage==7.6.1
dill==0.3.9
distlib==0.3.8
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
openai==1.35.10
packaging==24.1
pathspec==0.12.1
persian-tools==0.0.11
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.3.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
tomlkit==0.13.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
